// src/shared/api/models/getApiSalesList200SalesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiSalesList200SalesItemShortInvoiceId } from './getApiSalesList200SalesItemShortInvoiceId';
import type { GetApiSalesList200SalesItemParentId } from './getApiSalesList200SalesItemParentId';
import type { GetApiSalesList200SalesItemCustomerId } from './getApiSalesList200SalesItemCustomerId';
import type { GetApiSalesList200SalesItemCustomerName } from './getApiSalesList200SalesItemCustomerName';
import type { GetApiSalesList200SalesItemTips } from './getApiSalesList200SalesItemTips';
import type { GetApiSalesList200SalesItemSalesNote } from './getApiSalesList200SalesItemSalesNote';
import type { GetApiSalesList200SalesItemVoidId } from './getApiSalesList200SalesItemVoidId';
import type { GetApiSalesList200SalesItemSalesType } from './getApiSalesList200SalesItemSalesType';
import type { GetApiSalesList200SalesItemCreatedBy } from './getApiSalesList200SalesItemCreatedBy';
import type { GetApiSalesList200SalesItemUpdatedBy } from './getApiSalesList200SalesItemUpdatedBy';
import type { GetApiSalesList200SalesItemSellBy } from './getApiSalesList200SalesItemSellBy';
import type { GetApiSalesList200SalesItemSyncedWith } from './getApiSalesList200SalesItemSyncedWith';
import type { GetApiSalesList200SalesItemSyncId } from './getApiSalesList200SalesItemSyncId';
import type { GetApiSalesList200SalesItemDeletedAt } from './getApiSalesList200SalesItemDeletedAt';
import type { GetApiSalesList200SalesItemCustomer } from './getApiSalesList200SalesItemCustomer';

export type GetApiSalesList200SalesItem = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    invAutoNumber: number;
    salesDate: string;
    invoiceId: string;
    shortInvoiceId: GetApiSalesList200SalesItemShortInvoiceId;
    parentId?: GetApiSalesList200SalesItemParentId;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    outletId: number;
    customerId: GetApiSalesList200SalesItemCustomerId;
    customerName: GetApiSalesList200SalesItemCustomerName;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    subTotal: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    totalTax: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    totalDiscount: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    netTotal: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    tenderAmount: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    change: number;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    roundOff: number;
    tips: GetApiSalesList200SalesItemTips;
    salesNote: GetApiSalesList200SalesItemSalesNote;
    voidId: GetApiSalesList200SalesItemVoidId;
    salesType: GetApiSalesList200SalesItemSalesType;
    createdBy: GetApiSalesList200SalesItemCreatedBy;
    updatedBy: GetApiSalesList200SalesItemUpdatedBy;
    sellBy: GetApiSalesList200SalesItemSellBy;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    sourceId: number;
    syncedWith: GetApiSalesList200SalesItemSyncedWith;
    syncId: GetApiSalesList200SalesItemSyncId;
    createdAt: string;
    updatedAt: string;
    deletedAt?: GetApiSalesList200SalesItemDeletedAt;
    creditAmount?: number;
    customer?: GetApiSalesList200SalesItemCustomer;
    outstandingAmount?: number;
};