// src/shared/api/models/putApiCreditSaleIdBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiCreditSaleIdBodyPaymentItem } from './putApiCreditSaleIdBodyPaymentItem';
import type { PutApiCreditSaleIdBodyProductsItem } from './putApiCreditSaleIdBodyProductsItem';

export type PutApiCreditSaleIdBody = {
    payment?: PutApiCreditSaleIdBodyPaymentItem[];
    products?: PutApiCreditSaleIdBodyProductsItem[];
};